django/requirements.txt
django
django-bootstrap-v5
django-cors-headers
django-filter
djangorestframework
gunicorn
markdown
moment
psycopg2

# pytz==2017.3
